+++ pkg/epp/server/runserver.go
@@ -51,7 +51,7 @@ import (
 
 // ExtProcServerRunner provides methods to manage an external process server.
 type ExtProcServerRunner struct {
-	GrpcPort int
+	GrpcPort uint16
 	// GrpcListener is an optional pre-bound listener for the ext_proc server.
 	// When set, GrpcPort is ignored. Reserving the port in advance of this
 	// runnable starting closes the window in which another process can take a
@@ -100,7 +100,7 @@ func NewExtProcServerRunner(
 	priorityBandControlPlane contracts.PriorityBandControlPlane,
 ) *ExtProcServerRunner {
 	return &ExtProcServerRunner{
-		GrpcPort:                         int(opts.GRPCPort),
+		GrpcPort:                         opts.GRPCPort,
 		GKNN:                             gknn,
 		ControllerCfg:                    controllerCfg,
 		Datastore:                        ds,
